Move UI elements to their own package
New prompttoolkit abstractions will be added underneath
this package
